Template expansion in LHS of json path test
This change implements template substitution ($ENVIRON, $RESPONSE,
etc) in the left hand side of a response_json_paths test. This enables
testing in situations where responses may include a server side
generated identifier as the key in an object structure.

As things stand, this is the bare minimum implementation that assumes
that the test creator will be responsible for dealing with the
vargrancies of quoting that will be required in order to generate
legitimate JSON path expressions. For example if there is a uuid (with
hyphens) at $RESPONSE['$.id'] then this expression is likely to fail:

    $.nested.structure.$RESPONSE['$.id'].name: foobar

as it will evaluate to something like

    $.nested.structure.ADC8AAFC-D564-40D1-9724-7680D3C010C2.name: foobar

which _may_ be treated as an arithemtic expression by the json parser.
The test creator would need to do:

    $.nested.structure["$RESPONSE['$.id']"].name: foobar

or variants thereof to get a useful expression.

Docs to follow if this is considered an okay way to implement the
feature.

Fixes #170