Make _load_data_file public
Turns out this method is generally useful, so it should be public. The
old private method is maintained as is, with the public version simply
being a facade over it. In the future we may wish to consider extracting
it all to a utility function, but this is the minimal change to get some
usefulness. Tests are updated accordingly.