Add a tox target for testing against openstack placement
The openstack placement service is one of the larger consumers of gabbi.
This change adds a 'placement' tox target get retrieves the latest code
that provides placement and runs the relevant gabbi tests against it.

The target is added to the "allowed to fail" section travis.

This is considerably more complex than the gnocchi target (which does a
similar thing) because this gets the master tarball rather than cloning
the git repo. The clone takes far too long because the curren thost of
the placement service, nova, is very large and very old.