Correctly match regexes that start or end with slashes
Previously when the expected expression was a regex (denoted by
beginning and ending with a '/') we would obtain the actual regex by
stripping all leading and trailing slashes... and then stripping all
trailing slashes again to be sure, to be sure. That meant that if the
expected expression began or ended with one or more slashes, they would
be ignored when performing the regex match. Since '/' is the path
separator and paths are something we expect to compare often, this could
easily have led to tests succeeding where they should not.

In addition, the single-character expression '/' was treated as an
empty regex, when clearly it should be treated as a literal.