Don't treat data in substitutions as regular expressions
When the data we're trying to match includes a substitution (e.g. from
$HISTORY), we should never treat the substituted data as a regular
expression. Specifically:

1) If the user did not explicitly request regular expression matching,
the expected value is not longer treated like a regular expression even
if substitutions result in the expected value beginning and ending with
a '/'; and

2) If the user did explicitly request regular expression matching, any
substitutions are now escaped, meaning that they will match literally
within the regular expression and not be treated as regular expression
syntax themselves.