Add docs for use_prior_test
Because adding use_prior_test to the metadata section was going to
reflow the entire metadata table anyway, use this opportunity to change
the table to a list-table, which we either didn't know about when first
doing this, or didn't like it for some reason.

The new format is far more resilient to diffing, so a better long term
choice.

A FAQ entry is also added to indicate the availability of the feature.