Allow to load python object from yaml
It can be interesting to write custom object to compare values.

For example, I need to ensure an output is equal to .NAN

Because .NAN == .NAN always returns false. We currently can't compare it
with assert_equals().

With the unsafe yaml loader we can register a custom method to check
NAN, for example:

    class IsNAN(object):
        @classmethod
        def constructor(cls, loader, node):
            return cls()

        def __eq__(self, other):
            try:
                return math.isnan(other)
            except TypeError:
                return False

    yaml.add_constructor(u'!IsNAN', IsNAN.constructor)