Fix pytest handling of fixture-level skips
Due to an oversight when developing the pytest functionality, the
feature wherein if a fixture raised a SkipTest exception, all the
tests within the suite would be skipped did not work.

It turns out this is because several bad assumption were being made:

* That the start_fixture had access to the list of tests in
  that pytest will run. It doesn't. It has access to a list
  of tests prior to the tests being process by the pytester
  plugin. These are not the tests that will be run and
  manipulating them to be skipped will not work.

  Therefore in the ptester plugin pre-processing the list of
  real tests is added to the information the start_fixture is
  aware of.

* Calling addSkip on the pytest result class does not return.
  What needs to happen instead is that a skip marker is applied
  to the tests. Since we now have all the tests as described in
  the bullet above, it is possible to traverse them and mark
  them.

This was all tested against the live tests, which will skip
if GABBI_SKIP_NETWORK is true. In the process, it became clear
that there is no need to have different test loader files for
pytest versus unit tests. The load_tests and pytest_generate_tests
methods can live in the same file and will only be used by
the test loaders that want them.

So now all the YAML files are run by both test running, increasing
the coverage provided when using pytest.

Fixes #228