Raise GabbiFormatError when headers are None
It easy in YAML to create an entry which is normally a dict but
ends up with the value of None by making an empty entry:

    - name: foo
      GET: /
      request_headers:
      status: 200

This is something that gabbi has (accidentally) allowed, but should
not as it can lead to inscrutable problems throughout the tests.

So now that there's a convenient place to put it, raise
GabbiFormatError indicating the test where the problem is.

Fixes #236