Make it possible to use non-string $RESPONSE in data
Assuming a previous requested `GET /items` resulted in a response of:

```json
["a", "b", "c"]
```

We'd like a request like this (in a badly designed API) to work:

```yaml
name: delete some items named in the body
POST: /deletor
request_headers:
  content-type: application/json
data: $RESPONSE['$']
```

Prior to this change, this would not work. The subsitution would
be expanded to a list and sent down the http socket as a list, failing
because writing to a socket requires strings or bytes.

Now, when processing 'data' if, after template handling, we have
a complex data type, we'll attempt to dump it to the proper form
(from a list to a json representation of a list in this case).

This was the desired functionality all long but apparently no
one ever tried it.