Allow listing of tests with no host configured
When creating tests, if the host was empty (and no intercept was provided),
an AssertionError would be raised immediately. This meant that in the case
where no configuration information is available, authors of load_test()
functions had to choose between simply not loading any tests (or
suppressing the exception) and allowing test discovery to fail.

This change allows an empty string to be passed as the host, in which case
the TestCases will be generated (so tests can be listed) but the tests will
be skipped. This will show up in reporting, whereas excluding the tests
from the listing is silent.

If the host, url, and intercept are all None, this will continue to
generate an immediate AssertionError since it is likely due to the caller
simply not passing any of those arguments.

Signed-off-by: Zane Bitter <zbitter@redhat.com>