Rename YAMLHandler and update _register
This commit aims to add a class attribute called add_response_handling
to the ResponseHandler. It's purpose is to allow a ContentHandler to
exclude itself from the response_handler collection so that more than
one ContentHandler with the same content-type can run tests correctly.

YAMLHandler was renamed to something pretty long, I think we can
use YDL as an acronym for YAMLDiskLoading if we don't like the length.