Turn of uxsuccess validation in test harness
Because stestr now (correctly) treats uxsuccess as a signal
that the test suite should fail, we can no longer easily
confirm that generating a uxsuccess produces the expected
output, at least not without making some additional tests
outside the normal harness.

That's not worth doing. We are now able to accept and rely
on the fact that stestr does the right thing so we don't
have to have special tests for that.

Fixes: #258