Update testing to use python 3.7
Python 3.7 is common now, so make it the main test version
in tox.ini and update .travis.yml accordingly.

To have 3.7 we need to start using the xenial dist in travis, and
fall back to trusty for pypy.

In python 3.7 re.escape changes from escaping all non-alphanumeric,
to just special characters, requiring a change in test_history
for that version.

Trove classifiers are updated to reflect 3.7 support. Support for
3.4 is removed, not because gabbi doesn't work on 3.4 but because
it is not worth indicating support.