Make use of unsafe yaml forward compatible
The forthcoming 5.x release of PyYMAL changes how unsafe
yaml is managed [1]. The default for gabbi is to safe
load but the option has been present to unsafe load if
code-executing functionality was desired. The way to
achieve that in 5.x and beyond is more explicit than
prior versions.

In this change we adjust the unsafe loading to work with
both 3.x and the as yet unreleased 5.x.

Fixes #266

[1] https://msg.pyyaml.org/load