Enable a -q arg to gabbi-run
This will quiet the output from the test runner.

Because of the mess which is unittest test runners,
where output happens in both the result handler and
the runner but you can only control output in the
result handler, the easiest way to do this is to pass
a /dev/null stream to the runner and have the output
swallowed.