Pin pytest to <5.0.0
5.0.0 no longer wants you to use pytest.config. Gabbi relies on that.

While hunting for a different solution we'll use older versions.

Note also that 5.0.0 drops support for python 2.7 so when we start
using that we will need a solution that works in both 4.x and 5.x
so we can required 4.x for py 2.7 (which gabbi will continue to
support for a while).