Enable cast for ENVIRON
Because of the way re.sub works with functions, we need to do
casting outside of the repl function passed to re.sub. Those
functions are only allowed to return strings, and if they
don't the results are ignored.

Therefore we manage a 'cast' attribute on the test itself
which contains the value of any 'cast' group from the the
most recent matched regex. This is then used (via APPROVED_CASTS)
to do the actual casting to the proper value.

Interesting, this work points out some problems that already
existed in any casting done with replacers: the cast is done
on the entire message, not just the replaced bit, which means
for cases like

  foo$ENVIRON:int['12']bar

the results will be not quite as expected. I'll work this out
as this series of patches progresses, but for now, consider
it a known issue.