Drop support for py2 and remove testtools
The initial motivation for this change was removing the testtools
library. It has a cyclical dependency with the fixutures module
which makes installing gabbi in some environments (notably bazel)
difficult. Please to fix this problem in the testtools and fixtures
communities have gone unaddressed for _years_ so an effort was
undertaken to remove them.

That work exposed that much of what testtools was doing was masking
over differences in python3 and python3 unittests. Since it is 2020,
choosing to drop support for python2 is a legitimate choice. People
who still need support for python2 can use an earlier version of
gabbi.

To signal that change, the version number has been raised to 2.0.0.

tox, travis config and setup.cfg have been updated accordingly.