Automate container creation on release
This is based on the Dockerfile provided at #284 with some
tweaks.

Makefile is adjusted to allow easy experimentation, but the
primary work here uses github actions.

The container is built any time a tag that looks like a
semver is pushed. This is in alignment with the existing
release handling.