Make host_info_from_target work with modern urlparse (#292)
The modern urlparse has changed behavior for parsing
url-like strings that do not have a scheme. Initially
this was added in python 3.7.x but there were issues
with backwards compatibility. It came back with python
3.9.

The fix here addresses the problem by always adding
a scheme if one is not present. Existing styles of
telling gabbi where to go with requests continue
to work.

Python 3.9 is added to the testing configurations.

Fixes #277