Better test name reporting and skipping
A shortDescription() method is subclassed. The super
would return the doc string on the test method being
run. Setting __doc__ on that method is not allowed,
so shortDescription() is overridden to provide a
nice default:

  <the name of the test file>: <the name of the test>

Also, the `skip` test member may now use substititions.
This allows skipping to be controlled by environment
variables.