270 replacer casts (#275)
Sometimes we want the value of a replacer to be different
from its original type. This change adds explicit casting when
people need it for ENVIRON and RESPONSE.

Because of the way re.sub works with functions, we need to do
casting outside of the repl function passed to re.sub. Those
functions are only allowed to return strings, and if they
don't the results are ignored.

Therefore we manage a 'cast' attribute on the test itself
which contains the value of any 'cast' group from the the
most recent matched regex. This is then used (via APPROVED_CASTS)
to do the actual casting to the proper value.

If the replacement token is the whole of the message, then an
included :cast will be done. Otherwise there will be a RuntimeError.

This provides the precedent for the previous behavior on
$ENVIRON replacement: atoms can be cast because there is no
context to create a dependency problem.

See the docs changes for more details on how this works.

Fixes #270