Add yaml-based tests for host header and sni checking
The addition of server_hostname to the httpclient
PoolManager, without sufficient testing, has revealed
some issues:

* The minimum urllib3 required is too low. server_hostname
  was introduced in 1.24.x
* However, there is a bug [1] in PoolManager when mixing
  schemes in the same pool manager. This is being fixed
  so the minimum urllib3 will need to be higher still.

Tests are added here, and the minimum value for urllib3
will be set when a release is made.

Some of the tests are "live" meaning they require
network, and can be skipped via the live test
fixture if the GABBI_SKIP_NETWORK env variable is
set to "true".

[1] https://github.com/urllib3/urllib3/issues/2534

Fixes #307
Fixes #309