Fix test naming with pytest
It turns out that it was easy to create tests with duplicate
names in pytest, and given the global state used for managing
fixtures and test reporting, this could lead to obscure bugs
when running multiple test suites via pytest.

This problem is fixed here by making sure that the test name
is more likely to be unique. In order to do insure that there
is enough information available to "name" a test, py_test_generator
now requires that test_loader_name be set. Since py_test_generator
can only be called once per module, setting test_loader_name
to __name__ is sufficient for uniqueness.

To make parsing the test names for fixtures, the names are
more complex and use a `:`.