Add yaml-based tests for host header and sni checking (#310)
The addition of server_hostname to the httpclient
PoolManager, without sufficient testing, has revealed
some issues:

* The minimum urllib3 required is too low. server_hostname
  was introduced in 1.24.x
* However, there is a bug [1] in PoolManager when mixing
  schemes in the same pool manager. This is being fixed
  so the minimum urllib3 will need to be higher still.

Tests are added here, and the minimum value for urllib3
will be set when a release is made.

Some of the tests are "live" meaning they require
network, and can be skipped via the live test
fixture if the GABBI_SKIP_NETWORK env variable is
set to "true".

The added tests identified a significant problem with
test naming in pytest-handling that prevented them
from passing.

It turns out that it was easy to create tests with duplicate
names in pytest, and given the global state used for managing
fixtures and test reporting, this could lead to obscure bugs
when running multiple test suites via pytest.

This problem is fixed here by making sure that the test name
is more likely to be unique. In order to do insure that there
is enough information available to "name" a test, py_test_generator
now requires that test_loader_name be set. Since py_test_generator
can only be called once per module, setting test_loader_name
to __name__ is sufficient for uniqueness.

To make parsing the test names for fixtures, the names are
more complex and use a `:`.

[1] https://github.com/urllib3/urllib3/issues/2534

Fixes #307
Fixes #309