Handle double response formatters with mixed quotes (#340)
Prior to these changes a test entry like:

  url: $RESPONSE['$.json.path']$RESPONSE['$.json.otherpath']

would experience a parse error as the extracted jsonpath
would include portions of the 2nd $RESPONSE string.

This turns out to be because of a greedy regex.

What's interesting and confusing is that the regex, as
written, was supposed to be non-greedy already, but it
wasn't working correctly.

Further, trying to fix it while maintaining the use of
named groups and backreferences did not work.

The result is that we now do slighly more post-processing
on slight more complex regexes which might match on arg1
or arg2 (replacers) and case1 or case2 (history indicators)
instead of arg and case.

Even with this, there is now one prior test, in
json-left-side.yaml, that had to be changed in order to
work properly with the new regexes.

Unless we can come up with something clever, we may
need to accept this test change, which is probably fine
because the case was very corner.